delete unused code and renaming packages
