+++ src/main/java/app/AppBuilder.java
@@ -10,9 +10,9 @@ import data_access.InMemoryUserDataAccessObject;
 import entity.CommonUserFactory;
 import entity.UserFactory;
 import interface_adapter.ViewManagerModel;
-import interface_adapter.add_friend.ChatListController;
-import interface_adapter.add_friend.ChatListPresenter;
-import interface_adapter.add_friend.ChatListViewModel;
+import interface_adapter.chat_list.ChatListController;
+import interface_adapter.chat_list.ChatListPresenter;
+import interface_adapter.chat_list.ChatListViewModel;
 import interface_adapter.change_password.ChangePasswordController;
 import interface_adapter.change_password.ChangePasswordPresenter;
 import interface_adapter.change_password.LoggedInViewModel;
@@ -24,9 +24,9 @@ import interface_adapter.logout.LogoutPresenter;
 import interface_adapter.signup.SignupController;
 import interface_adapter.signup.SignupPresenter;
 import interface_adapter.signup.SignupViewModel;
-import use_case.ChatList.ChatListInputBoundary;
-import use_case.ChatList.ChatListManager;
-import use_case.ChatList.ChatListOutputBoundary;
+import use_case.chat_list.ChatListInputBoundary;
+import use_case.chat_list.ChatListManager;
+import use_case.chat_list.ChatListOutputBoundary;
 import use_case.change_password.ChangePasswordInputBoundary;
 import use_case.change_password.ChangePasswordInteractor;
 import use_case.change_password.ChangePasswordOutputBoundary;
