+++ src/main/java/interface_adapter/login/LoginPresenter.java
@@ -1,10 +1,8 @@
 package interface_adapter.login;
 
 import interface_adapter.ViewManagerModel;
-import interface_adapter.add_friend.ChatListState;
-import interface_adapter.add_friend.ChatListViewModel;
-import interface_adapter.change_password.LoggedInState;
-import interface_adapter.change_password.LoggedInViewModel;
+import interface_adapter.chat_list.ChatListState;
+import interface_adapter.chat_list.ChatListViewModel;
 import use_case.login.LoginOutputBoundary;
 import use_case.login.LoginOutputData;
