+++ src/main/java/use_case/chat_list/ChatListOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.ChatList;
+package use_case.chat_list;
 
 import entity.ChatEntry;
